requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
aiosqlite==0.19.0
sqlalchemy==2.0.23
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
pytz==2023.3
cachetools==5.3.2